third_party/blink/web_tests/fast/css/cssimagesetvalue-crash.html

<html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>

<div id="target"></div>

<script>
function testImageSetSerialization(isPrefixed) {
  const prefix = isPrefixed? "-webkit-" : "";

  target.style.backgroundImage = '';
  target.style.backgroundImage = `${prefix}image-set(url(test1.png) 1x, url(test2.png) 2x)`;

  const url = location.href;
  const path = url.substring(0, url.lastIndexOf('/')+1);

  const expected = `${prefix}image-set(url("${path}test1.png") 1x, url("${path}test2.png") 2x)`;

  assert_equals(
    getComputedStyle(target).backgroundImage,
    expected
  );

  assert_equals(
    target.style.backgroundImage,
    `${prefix}image-set(url("test1.png") 1x, url("test2.png") 2x)`
  );
}

test(function() {
  testImageSetSerialization(false);
}, 'Testing serialization of image-set');

test(function() {
  testImageSetSerialization(true);
}, 'Testing serialization of -webkit-image-set');
</script>
</html>
